Add scripts, configuration files and documentation for running live
The easiest way get to up and running to develop and run bots is using
Docker. Using Docker will ensure that all dependencies are available at
the expected versions and allow easy upgrades if necessary. Also, this
minimizes the effect of the choice of platform (Windows or UNIX-like) on
the bot, configuration and documentation.

Additionally, instructions are provided for people not wanting or able
to use Docker. These instructions recommend the use of a virtual
environment, which can be created in one go with the create_venv.sh
script.

Since ta-lib is installed in the Docker image, this commit also
re-enables the TA-Lib Python dependency.